# use LFU as crs