--- a/crs/crs.go
+++ b/crs/crs.go
@@ -1,0 +1,7 @@
+// crs Cache Replacement Strategy
+
+package crs
+
+import "github.com/andersfylling/disgord/depalias"
+
+type Snowflake = depalias.Snowflake