--- a/crs/run-benchmarks.sh
+++ b/crs/run-benchmarks.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+go test -bench=BenchmarkCRS -count 10 >results-native.txt
+
+benchstat results-interface.txt results-native.txt >perf-diff.txt