# Description

Drop the cache replacement interface and all the other overhead and use LFU. This is the first step in implementing a TLFU cache system.

Affects #180 

## Benchmarks

> CRS/get-lfu-8 is ignored as the compiler seem to have optimized away something when using the interface version.

Current LFU:

|                    name | time/op |
|------------------------:|---------|
| CRS/set-lfu-unlimited-8 |  132ns ± 3%  |
|   CRS/set-lfu-limited-8 |    8.39µs ± 1%    |
|        CRS/update-lfu-8 |    133ns ± 2%     |
|           ~CRS/get-lfu-8~ |    ~0.86ns ± 4%~     |

This PR:

|                    name | time/op |
|------------------------:|---------|
| CRS/set-lfu-unlimited-8 |  105ns ± 1%  |
|   CRS/set-lfu-limited-8 |    8.19µs ± 3%    |
|        CRS/update-lfu-8 |    104ns ± 1%    |
|           CRS/get-lfu-8 |    38.1ns ± 2%     |

Diff:
```
name                     old time/op  new time/op   delta
CRS/set-lfu-unlimited-8   132ns ± 3%    105ns ± 1%    -20.94%  (p=0.000 n=8+8)
CRS/set-lfu-limited-8    8.39µs ± 1%   8.19µs ± 3%     -2.40%  (p=0.002 n=10+10)
CRS/update-lfu-8          133ns ± 2%    104ns ± 1%    -21.73%  (p=0.000 n=10+9)
```
|                    name | old time/op | new time/op | delta |
|------------------------:|-------------|-------------|-------|
| CRS/set-lfu-unlimited-8 |    132ns ± 3%  |    -20.94%  |-20.94%  (p=0.000 n=8+8)|
|   CRS/set-lfu-limited-8 |   8.39µs ± 1%  |    -2.40%   |-2.40%  (p=0.002 n=10+10)|
|        CRS/update-lfu-8 |  133ns ± 2%  |     -21.73% |-21.73%  (p=0.000 n=10+9)|
|           CRS/get-lfu-8 |      ~0.86ns ± 4%~       |      38.10ns ± 2%       |  ~+4309.72%  (p=0.000 n=10+9)~     |

# Checklist:

- [x] I ran `go generate`
- [x] I have performed a self-review of my own code (remember to run `go fmt ./...`)
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules
- [x] Added benchmarks if this is a performant required component (potential bottlenecks)